Move design-to-fail samples to separate package to avoid build failures
